backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.3
sqlalchemy==2.0.36
python-telegram-bot==21.10
python-dotenv==1.0.1
pydantic==2.10.6